requirements.txt

fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-multipart
langchain
langchain-community
chromadb
sentence-transformers
transformers
torch
datasets
pandas
PyPDF2
python-docx
jose
